# JXL-in-TIFF: implement support NBITS ==12 for UInt16 data type (fixes #6409)