This requires libjxl post 0.8.1 / https://github.com/libjxl/libjxl/pull/1812