@tbonfort 2023-02-08
Thanks Even, I can confirm that this implementation produces similar results to my first-draft hard-coded patch from #6409.

However this double/useless bit packing/unpacking makes me sad. I was hoping that it was just me worrying about nothing, but running some timings today shows a significant slowdown:

On a 1024x1024 tile:

- using this PR's implementation (i.e. NBITS=12):
  - compression: ~0.27sec
  - decompression: ~0.24sec

-  using my dirty implementation (NBITS=16, only the jxl codestream is marked as 12bits):
  - compression: ~0.14sec
  - decompression: ~0.10 sec

I'm not sure what the way forward is here... I'm weary of introducing a sub-par implementation (performance wise) that might bring in backwards compatibility issues later on if/when we implement a more performant one, but I'm also weary of introducing a more performant one that has known forward compatibility issues. Maybe we should leave this on hold until we find an acceptable solution?