@rouault 2023-02-08
How do you manage to get such small compression/decompression times... ?

```
$ gdal_translate autotest/gdrivers/data/small_world.tif 1024.tif -outsize 1024 1024 -r bilinear -scale 0 255 0 4095 -co nbits=12 -ot uint16
$ time gdal_translate 1024.tif 1024_jxl.tif -co compress=jxl -co jxl_distance=1 -co jxl_lossless=no
real	0m2,031s
user	0m1,993s
$ time gdal_translate 1024.tif 1024_jxl.tif -co compress=jxl
real	0m1,587s
user	0m1,562s
$ time gdal_translate 1024.tif 1024_jxl.tif -co compress=jxl -co tiled=yes
real	0m1,159s
user	0m1,118s
$ time gdal_translate 1024.tif 1024_jxl.tif -co compress=jxl -co tiled=yes -co jxl_distance=1 -co jxl_lossless=no
real	0m0,346s
user	0m0,321s

```
I've checked that both my GDAL and libjxl are built in Release mode