@tbonfort 2023-02-08
> How do you manage to get such small compression/decompression times... ?

I use JXL_EFFORT=3 and a single tile:

time gdal_translate -co TILED=YES -co BLOCKXSIZE=1024 -co BLOCKYSIZE=1024 -co COMPRESS=JXL -co JXL_LOSSLESS=FALSE -co JXL_EFFORT=3 -co NBITS=12 tile-in.tif tile-jxl12.tif