@rouault 2023-02-08
ok, JXL_EFFORT=3 was the trick:
```
$ time gdal_translate -co TILED=YES -co BLOCKXSIZE=1024 -co BLOCKYSIZE=1024 -co COMPRESS=JXL -co JXL_LOSSLESS=FALSE -co JXL_EFFORT=3 -co NBITS=12 1024.tif 1024_jxl.tif

real	0m0,197s
user	0m0,161s
```