@rouault 2023-02-08
I've just committed a big performance boost by borrowing the optimized nbits==12 bit packing/unpacking from the GTiff reader, instead of relying on the generic code. I've also only enabled nbits==12 for uint16 nbits != 16

Timing before:
```
$ hyperfine --warmup 2 'gdal_translate -co TILED=YES -co BLOCKXSIZE=1024 -co BLOCKYSIZE=1024 -co COMPRESS=JXL -co JXL_LOSSLESS=FALSE -co JXL_EFFORT=3 -co NBITS=12 1024.tif 1024_jxl.tif -q'
Benchmark #1: gdal_translate -co TILED=YES -co BLOCKXSIZE=1024 -co BLOCKYSIZE=1024 -co COMPRESS=JXL -co JXL_LOSSLESS=FALSE -co JXL_EFFORT=3 -co NBITS=12 1024.tif 1024_jxl.tif -q
  Time (mean ± σ):     184.2 ms ±   1.5 ms    [User: 142.8 ms, System: 40.8 ms]
  Range (min … max):   182.4 ms … 189.1 ms
 
$ hyperfine --warmup 2 'gdal_translate 1024_jxl.tif tmp.tif -scale 0 4095 0 255 -ot byte'
Benchmark #1: gdal_translate 1024_jxl.tif tmp.tif -scale 0 4095 0 255 -ot byte
  Time (mean ± σ):     180.7 ms ±   0.9 ms    [User: 155.7 ms, System: 24.2 ms]
  Range (min … max):   178.6 ms … 181.9 ms
```

Timing after:
```
$ hyperfine --warmup 2 'gdal_translate -co TILED=YES -co BLOCKXSIZE=1024 -co BLOCKYSIZE=1024 -co COMPRESS=JXL -co JXL_LOSSLESS=FALSE -co JXL_EFFORT=3 -co NBITS=12 1024.tif 1024_jxl.tif -q'
Benchmark #1: gdal_translate -co TILED=YES -co BLOCKXSIZE=1024 -co BLOCKYSIZE=1024 -co COMPRESS=JXL -co JXL_LOSSLESS=FALSE -co JXL_EFFORT=3 -co NBITS=12 1024.tif 1024_jxl.tif -q
  Time (mean ± σ):     110.7 ms ±   1.0 ms    [User: 72.3 ms, System: 37.8 ms]
  Range (min … max):   108.7 ms … 112.9 ms
 
$ hyperfine --warmup 2 'gdal_translate 1024_jxl.tif tmp.tif -scale 0 4095 0 255 -ot byte'
Benchmark #1: gdal_translate 1024_jxl.tif tmp.tif -scale 0 4095 0 255 -ot byte
  Time (mean ± σ):      94.0 ms ±   0.7 ms    [User: 68.4 ms, System: 25.1 ms]
  Range (min … max):    93.0 ms …  96.4 ms
 ```