@tbonfort 2023-02-09
Thanks Even,
I'm still wondering if it is desirable to add this functionality now, and have to bear with potential backwards incompatibilities later. Understandably, (at least from what I understand) jxl's lossy mode needs to know what uint16 value corresponds to "black"  and what value corresponds to "white", for each band, in order to apply its perceptually driven compression. Those two values cannot be expressed by a single number of bits, so even if we were able to support an arbitrary number NBITS instead of only 12, that would not be enough.

My feeling is that we should close this issue for now. For the future:
- we should not rely on NBITS for setting the input data range
- we need a libjxl api that allows setting the black and white points, maybe for each band independently (I haven't checked if this api already exists or not)
- the tiff spec does have some tags to express the dynamic range with MinSampleValue, MaxSampleValue, but they are marked as informative only. I don't see how a gdal user could express those without using creation options, which I understand are undesirable to extend.

Happy to discuss this more with anyone interested if needed.