@rouault 2023-02-09
For the black and white points, it could potentially be conveyed to libtiff through the https://www.awaresystems.be/imaging/tiff/tifftags/referenceblackwhite.html tag. But indeed we' need a GTiff creation option for that. There's no particular problem in adding one to the many existing ones...
I also agree that the NBITS=12 case is probably not optimum as it. Probably we should use a dedicated TIFF tag specific to the JXL codec to mean "codec bitdepth, but not libtiff bitdepth", and that would probably need to be a real TIFF tag encoded in the file, not just an ephemeral pseudo-tag.
Closing